+++ tests/providers/apache/hive/operators/test_hive.py
@@ -26,7 +26,6 @@ from airflow import DAG
 from airflow.configuration import conf
 from airflow.exceptions import AirflowSensorTimeout
 from airflow.models import TaskInstance
-from airflow.operators.presto_to_mysql import PrestoToMySqlTransfer
 from airflow.providers.apache.hdfs.sensors.hdfs import HdfsSensor
 from airflow.providers.apache.hdfs.sensors.web_hdfs import WebHdfsSensor
 from airflow.providers.apache.hive.operators.hive import HiveOperator
@@ -36,6 +35,7 @@ from airflow.providers.apache.hive.operators.hive_to_samba import Hive2SambaOper
 from airflow.providers.apache.hive.sensors.hive_partition import HivePartitionSensor
 from airflow.providers.apache.hive.sensors.metastore_partition import MetastorePartitionSensor
 from airflow.providers.apache.hive.sensors.named_hive_partition import NamedHivePartitionSensor
+from airflow.providers.mysql.operators.presto_to_mysql import PrestoToMySqlTransfer
 from airflow.providers.presto.operators.presto_check import PrestoCheckOperator
 from airflow.sensors.sql_sensor import SqlSensor
 from airflow.utils import timezone
